+++ src/dune_rules/cxx_flags.ml
@@ -23,19 +23,6 @@ let base_cxx_flags ~for_ cc =
 
 let preprocessed_filename = "ccomp"
 
-let ccomp_type build_dir =
-  let open Action_builder.O in
-  let filepath =
-    Path.Build.(
-      relative (relative build_dir ".dune/ccomp") preprocessed_filename)
-  in
-  let+ ccomp = Action_builder.contents (Path.build filepath) in
-  match String.trim ccomp with
-  | "clang" -> Clang
-  | "gcc" -> Gcc
-  | "msvc" -> Msvc
-  | s -> Other s
-
 let check_warn = function
   | Other s ->
     User_warning.emit
@@ -47,8 +34,25 @@ let check_warn = function
       ]
   | _ -> ()
 
+let ccomp_type ctx =
+  let build_dir = ctx.Context.build_dir in
+  let open Action_builder.O in
+  let filepath =
+    Path.Build.(
+      relative (relative build_dir ".dune/ccomp") preprocessed_filename)
+  in
+  let+ ccomp = Action_builder.contents (Path.build filepath) in
+  let ccomp_type =
+    match String.trim ccomp with
+    | "clang" -> Clang
+    | "gcc" -> Gcc
+    | "msvc" -> Msvc
+    | s -> Other s
+  in
+  check_warn ccomp_type;
+  ccomp_type
+
 let get_flags ~for_ ctx =
   let open Action_builder.O in
-  let+ ccomp_type = ccomp_type ctx.Context.build_dir in
-  check_warn ccomp_type;
+  let+ ccomp_type = ccomp_type ctx in
   base_cxx_flags ~for_ ccomp_type